pyproject.toml
[tool.poetry]
name = "libertai-utils"
version = "0.0.16"
description = "LibertAI internal types and functions"
authors = ["LibertAI.io team <hello@libertai.io>"]
readme = "README.md"
homepage = "https://libertai.io"
repository = "https://github.com/LibertAI/libertai-utils"
documentation = "https://docs.libertai.io"
classifiers = [
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[tool.poetry.dependencies]
python = "^3.11"
web3 = "^7.10.0"
eciespy = "^0.4.2"
pydantic = "^2.10.6"
solders = ">=0.22.0,<1.0.0"
pynacl = "^1.5.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.11.1"
ruff = "^0.6.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
